Revised time evolution view, including the time evolution of state variables and the time evolution of noise around fixed points:
A list of state variables can be supplied as the first argument in numSimStateVar that contains all state variables which will be shown in the figure. If no list is supplied numSimStateVar plots all time-dependent state variables defined in the model. The remaining keyword arguments are optional.

To display noise numSimNoiseCorrelations can be used as before with the difference that there are only optional keyword arguments. State variables do not need to be given as input - all noise-noise correlations will be plotted automatically.

MuMoTtest_NoiseFixedPoints.ipynb has been updated to show how to make use of the new functionality.