Implemented plotProportion in numSimStateVar() via keyword argument
Default is that proportions are plotted.
If plotProportion=False is used in numSimStateVar() then
the concentrations (proportions) are multiplied by the value
on the System size slider or the systemSize defined via params
times a scaling parameter that makes the system size input equal to
the system size at t=0.
If systemSize is defined via paramsparams the System size slider
is deactivated automatically.

If proportions are plotted the legends show Phi_{A}, for example,
where A is a reactant
and if absolute numbers are plotted reactant A appears in the legend
to label the curve (this applies to every reactant with arbitrary label
that replaces A mentioned above).