Create new Integrate method (a MuMoTmodel method for numerical integration) which will replace numSimStateVar. Create new controller class MuMoTtimeEvolutionController. Create new view class MuMoTtimeEvolutionViewNew which will replace MuMoTtimeEvolutionView.
This commit solves issues #80, #93 and #97.

Initial conditions can now be set in the advanced options tab.
The values are constrained if the new keyword 'conserved'
is True. By default 'conserved'=False, except when
the system size has been set automatically after
applying substitute, then it is changed to True.

Maximum integration time can be set in the advanced options too.

Plot proportions can be chosen by ticking the correponding check-box in the advanced options tab.

Some minor edits in this commit:
Reduced tick-label size and legend-fontsize for a better appearance.