Solve LaTeX rendering problems when displaying mathematical formulas with Greek letters (when indices are involved) and with double indices.
This should be a global solution which can be applied anywhere in MuMoT.py, such as axes labels, display of equations, controller display. This is necessary as process_sympy() and the sympy function latex() have problems with Greek letters when indices are involved.

Created two lists containing the greek letters available (defined at the top of MuMoT.py).
Created three new functions which ensure proper handling of double indices and Greek letters:

_doubleUnderscorify(s), _greekPrependify(s), _greekReplace(s, sub, repl)
(all functions act on strings and are defined at the end of MuMoT.py)

In this commit these function have been applied in various places. Exceptions are the visualise method and ssa and multiagent-related classes and functions.