Change symbol in van Kampen expansion from V to \overline{V}.
The volume used in the van Kampen expansion is only displayed when
using the showVanKampenExpansion() method. Apart from that it is always
absorbed in \Phi used to express concentrations of reactants (e.g. \Phi_{A}=A/\overline{A}).

Internally in various functions we still use V to refer to \overline{V} by defining
sympy symbols according to V = Symbol('\overline{V}').

There is no interference when using V as state variable or rate
and the volume V used in the system size expansion and related functions.